change: remove demo tab and add mantime table
