+++ package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@fontsource/inter": "^5.2.5",
     "@lottiefiles/dotlottie-react": "^0.13.3",
+    "@mantine/core": "^8.0.1",
+    "@mantine/hooks": "^8.0.1",
     "@reduxjs/toolkit": "^2.7.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/user-event": "^14.6.1",
@@ -57,6 +59,9 @@
     "ajv": "^8.17.1",
     "ajv-keywords": "^5.1.0",
     "netlify-cli": "^20.0.2",
+    "postcss": "^8.5.3",
+    "postcss-preset-mantine": "^1.17.0",
+    "postcss-simple-vars": "^7.0.1",
     "typescript": "^4.9.5"
   }
 }
