+++ src/navigation/RouteNames.ts
@@ -6,7 +6,6 @@ const routeNames = {
   dashboard: '/app/dashboard',
   orders: '/app/orders',
   settings: '/app/settings',
-  demo: '/app/demo',
   noWhere: '*',
 };
