+++ src/navigation/Router.tsx
@@ -6,8 +6,7 @@ import SignIn from '../view/pages/auth/Signin';
 // Import lazy-loaded components
 const DashboardLayout = lazy(async () => await import('../view/wrappers/dashboard/DashboardLayout'));
 const Dashboard = lazy(async () => await import('../view/pages/Dashboard/Dashboard'));
-const Orders = lazy(async () => await import('../view/pages/Orders'));
-const Demo = lazy(async () => await import('../view/pages/Demo'));
+const Orders = lazy(async () => await import('../view/pages/Orders/orders'));
 const Settings = lazy(async () => await import('../view/pages/Settings'));
 
 const Router = () => {
@@ -20,7 +19,6 @@ const Router = () => {
         { path: routeNames.dashboard, element: <Dashboard /> },
         { path: routeNames.orders, element: <Orders /> },
         { path: routeNames.settings, element: <Settings />},
-        { path: routeNames.demo, element: <Demo />},
       ],
     },
     {
